+++ src/Eloquent/AbstractAdapter.php
@@ -36,6 +36,7 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Collection;
 use Neomerx\JsonApi\Contracts\Encoder\Parameters\EncodingParametersInterface;
 use Neomerx\JsonApi\Contracts\Encoder\Parameters\SortParameterInterface;
+use Neomerx\JsonApi\Encoder\Parameters\EncodingParameters;
 
 /**
  * Class EloquentAdapter
@@ -220,7 +221,9 @@ abstract class AbstractAdapter extends AbstractResourceAdaptor
             throw new RuntimeException('Paging parameters exist but paging is not supported.');
         }
 
-        return $pagination->isEmpty() ? $this->all($query) : $this->paginate($query, $parameters);
+        return $pagination->isEmpty() ?
+            $this->all($query) :
+            $this->paginate($query, $this->normalizeParameters($parameters, $pagination));
     }
 
     /**
@@ -769,6 +772,28 @@ abstract class AbstractAdapter extends AbstractResourceAdaptor
         list($one, $two, $caller) = debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS, 3);
 
         return $caller['function'];
+    }
+
+    /**
+     * Normalize parameters for pagination.
+     *
+     * This is a temporary solution for Issue #131 in the v0.11.x series.
+     *
+     * @param EncodingParametersInterface $parameters
+     * @param Collection $extractedPagination
+     * @return EncodingParameters
+     * @see https://github.com/cloudcreativity/laravel-json-api/issues/131
+     */
+    private function normalizeParameters(EncodingParametersInterface $parameters, Collection $extractedPagination)
+    {
+        return new EncodingParameters(
+            $parameters->getIncludePaths(),
+            $parameters->getFieldSets(),
+            $parameters->getSortParameters(),
+            $extractedPagination->all(),
+            $parameters->getFilteringParameters(),
+            $parameters->getUnrecognizedParameters()
+        );
     }
 
 }